Address @mmcgr's suggestions
