+++ src/Interpreter.cpp
@@ -26,13 +26,13 @@
 #include "Logger.h"
 #include "ParallelUtils.h"
 #include "ProfileEvent.h"
-#include "RamExistenceCheck.h"
+#include "RamExistenceCheckAnalysis.h"
 #include "RamIndexScanKeys.h"
 #include "RamNode.h"
 #include "RamOperation.h"
 #include "RamOperationDepth.h"
 #include "RamProgram.h"
-#include "RamProvenanceExistenceCheck.h"
+#include "RamProvenanceExistenceCheckAnalysis.h"
 #include "RamValue.h"
 #include "RamVisitor.h"
 #include "ReadStream.h"
