+++ src/RamOperation.h
@@ -695,7 +695,7 @@ public:
     RamReturn* clone() const override {
         std::vector<std::unique_ptr<RamValue>> newValues;
         for (auto& cur : values) {
-            newValues.push_back(std::unique_ptr<RamValue>(cur->clone()));
+            newValues.emplace_back(cur->clone());
         }
         return new RamReturn(std::move(newValues));
     }
