+++ src/Synthesiser.cpp
@@ -21,12 +21,12 @@
 #include "IODirectives.h"
 #include "IndexSetAnalysis.h"
 #include "RamCondition.h"
-#include "RamExistenceCheck.h"
+#include "RamExistenceCheckAnalysis.h"
 #include "RamIndexScanKeys.h"
 #include "RamNode.h"
 #include "RamOperation.h"
 #include "RamProgram.h"
-#include "RamProvenanceExistenceCheck.h"
+#include "RamProvenanceExistenceCheckAnalysis.h"
 #include "RamRelation.h"
 #include "RamTranslationUnit.h"
 #include "RamValue.h"
